feat(nutrition): store a food's calories and macros with decimals
A label prints 274.4 kcal and 6.7 g of protein; a food now keeps them
as printed instead of rounding to whole numbers. A meal logged from it
still rounds once, from the exact figure. The columns stay as they are,
since SQLite keeps a decimal written to them; the archive's food figures
widen from integer to number, so existing backups still import.

+++ lib/backend/engines/food_portion.dart
@@ -76,6 +76,6 @@ class FoodPortion {
       ? '${formatAmount(amount)} ${food.servingUnit.label}'
       : '${formatAmount(servings)} ${ServingUnit.serving.label}';
 
-  int? _scaled(int? perServing) =>
+  int? _scaled(double? perServing) =>
       perServing == null ? null : (perServing * servings).round();
 }
